==> Scripts/2_Stockholm_analyses/4_transformations_and_normalisations.R <==
#Now it is time to normalise and transform the gex data, and transform the 
#protein data. 
library(SingleCellExperiment)
library(scater)
library(scran)
library(bluster)
library(DepecheR)
library(ggplot2)
library(flowSpecs)
library(flowCore)
library(ggpubr)
fullSce <- readRDS("../External/Stockholm/Data/SCE_versions/1_full_SCE.rds")

#Do do this, we now need to cluster the cells according to their transcriptomic
#profiles. These clusters will however not be used downstream.  
set.seed(0101010)
kgraph.clusters <- clusterCells(fullSce, use.dimred="totalVI_lat_rep",
                                BLUSPARAM=TwoStepParam(
                                    first=KmeansParam(centers=500),
                                    second=NNGraphParam(k=10)
                                )
)
table(kgraph.clusters)

fullSce$Clusters <- kgraph.clusters

dColorPlot(fullSce$Clusters, xYData = reducedDim(fullSce, "UMAP"), 
           plotName = "Clusters_for_sumFactorCalc", plotDir = "Diagnostics/Stockholm")

#This looks reasonable. 
deconv.fullSce <- calculateSumFactors(fullSce, cluster=kgraph.clusters)
summary(deconv.fullSce)
#And these are saved, as they take so much time to generate. 
dir.create("../External/Stockholm/Data/SCE_auxiliaries")
saveRDS(deconv.fullSce, "../External/Stockholm/Data/SCE_auxiliaries/calculated_sum_factors.rds")
#deconv.fullSce <- readRDS("../External/Stockholm/Data/SCE_auxiliaries/calculated_sum_factors.rds")
fullSce <- logNormCounts(fullSce, size.factors = deconv.fullSce)

#Here, we after some emprical testing decided to go for an intermediate cofactor
#of 50, as it seems to make sense for the proteins in general. 
normcounts(altExp(fullSce)) <- asinh(counts(altExp(fullSce))/50)

#Here, we also run another diagnostic test, namely checking if there are spurious
#correlations between the proteins. Here, only the Kotliarov data is shown. 
ctrlRows <- which(fullSce$group == "Ctrl")
normAdtCountsCtrl <- t(normcounts(altExp(fullSce)))[ctrlRows,]
oneVsAllPlot(flowFrame(normAdtCountsCtrl), 
             dirName = "Diagnostics/Stockholm/Estimation_controls/All_vs_all_plots_after_totalVi")

#We also run this for the real protein data for the original Kotliarov dataset alone
adtCounts <- read.csv("../External/Stockholm/Data/Anndata_input/adt_counts.csv", row.names = 1)
adtCountsFull <- asinh(adtCounts/50)
adtCountsCtrl <- adtCountsFull[ctrlRows,]

oneVsAllPlot(flowFrame(as.matrix(adtCountsCtrl)), 
             dirName = "Diagnostics/Stockholm/Estimation_controls/All_vs_all_plots_before_totalVi")

#In an other analysis (sheet 4b), the differences here are quantified and visualised
#more directly. 
#This sadly shows that there is a lot of spuriosity going on. We will therefore 
#need to go about this in a flow-cytometry-associated way, by introducing a correction
#matrix. We will optimise this square matrix with a diagonal of 1s, so that the
#squware root of the sum of absolute residuals, or root mean square error (RMSE),
#between the original matrix and the
#first normalised and now corrected matrix of protein values for the Kotliarov data is minimised. 

#For the below comparison, we of course need to remove the two variables that are
#not present in the dataset, namely CD14 and CD45 and we will also make sure that
#the data is on similar scales. 
excluded <- c("CD014_PROT", "CD045_PROT")
normAdtCountsCtrl[,which(colnames(normAdtCountsCtrl) %in% excluded)] <- 0
identical(colnames(adtCountsCtrl), colnames(normAdtCountsCtrl)) #TRUE
identical(rownames(adtCountsCtrl), rownames(normAdtCountsCtrl)) #TRUE

#Now, we are going to select a subset of cells, that is as diverse as possible. 
#For this reason, we pre-cluster the cells and then take an equal number from 
#each cluster. 
set.seed(11122)
protClusters <- clusterCells(SingleCellExperiment(assays = list(logcounts = t(adtCountsCtrl))), 
                                assay.type= "logcounts",
                                BLUSPARAM=TwoStepParam(
                                    first=KmeansParam(centers=5000),
                                    second=NNGraphParam(k=5)
                                )
)
table(protClusters)
#protClusters
#1    2    3    4    5    6    7    8    9   10   11   12   13   14   15   16   17   18   19   20   21   22   23 
#2262 5008 4612 6364 1602 2831 4253 8603 6562 1185 4498 2351  107   99  967  604 1285  747  183  410  287   38   66 
#24 
#64 

dColorPlot(protClusters, xYData = reducedDim(fullSce[,ctrlRows], "UMAP"), 
           plotName = "Clusters_for_protein_signal_correction", plotDir = "Diagnostics/Estimation_controls")

write.csv(protClusters, "Data/Stockholm/Prot_clusters_for_protein_signal_correction.csv")

#As we have 24 clusters and the smallest one is containing 38 events, we will take 38 from each and 
#iterate 1000 times, in reality giving increased weight to the uncommon events.  
corrMatList <- bplapply(1001:2000, function(i){
    print(i)
    set.seed(i)
    rowSample <- unlist(lapply(unique(protClusters), function(x){
        sample(which(protClusters == x), 38)
    }))
    rawDatSubsetDf <- as.data.frame(adtCountsCtrl[rowSample,])
    normDatSubsetDf <- as.data.frame(normAdtCountsCtrl[rowSample,])
    model <- lm(as.matrix(adtCountsCtrl[rowSample,]) ~ as.matrix(normAdtCountsCtrl[rowSample,]) - 1)
    k <- coef(model)
})
corrMatBig <- do.call("rbind", lapply(corrMatList, function(x){
    as.vector(x)
}))
corMatFinal <- matrix(colMeans(corrMatBig), 64, 64)
corMatFinal[which(is.na(corMatFinal))] <- 0

# Use matrix multiplication to transform x into the likeness of y
transformed_x <- normAdtCountsCtrl %*% corMatFinal
colnames(transformed_x) <- colnames(normAdtCountsCtrl)

oneVsAllPlot(flowFrame(transformed_x), 
             dirName = "Diagnostics/Estimation_controls/All_vs_all_plots_after_totalVi_after_lin_corr")

#This works like a charm. Now, we will save this matrix. 
dir.create("Data/Stockholm/totalVI_protein_corrections")
write.csv(corMatFinal, "Data/Stockholm/totalVI_protein_corrections/Correction_matrix_least_squares_1000_bootstraps.csv")
#corMatFinal <- 
#  as.matrix(read.csv("Data/Stockholm/totalVI_protein_corrections/Correction_matrix_least_squares_1000_bootstraps.csv", 
#                     row.names = 1))

#So with that, we will add this back to the data
datRes <- t(normcounts(altExp(fullSce)))%*%corMatFinal
colnames(datRes) <- colnames(normAdtCountsCtrl)
oneVsAllPlot(flowFrame(datRes), 
             dirName = "Diagnostics/Estimation_controls/All_vs_all_plots_after_totalVi_after_lin_corr_all_cells")

#Before adding this back, we will include the CD45 and CD14 signal from the normalised dataset
#as these are lacking from this version. 
datRes[,"CD014_PROT"] <- normcounts(altExp(fullSce))["CD014_PROT",]
datRes[,"CD045_PROT"] <- normcounts(altExp(fullSce))["CD045_PROT",]
logcounts(altExp(fullSce)) <- t(datRes)

#And we also print all the proteins. 
dColorPlot(t(normcounts(altExp(fullSce))), xYData = reducedDim(fullSce, "UMAP"), 
           plotDir = "Diagnostics/Estimation_controls/Normalised_proteins")
datResNoNA <- datRes[,-which(colnames(datRes) %in% excluded)]
dColorPlot(datResNoNA, xYData = reducedDim(fullSce, "UMAP"), 
           plotDir = "Diagnostics/Estimation_controls/Normalised_corrected_proteins")

#Now, some sanity checks before saving. 
#At this stage, it is also prudent to make sure that the protein expression
#in the original datasets correspond well to this protein expression. 
#We will for this purpose import the original protein expression data.
#We will for this analysis perform separate analyses for the patient and the
#control data. 

rawProtDatPat <- adtCountsFull[which(fullSce$group == "Pat"),]
patColSums <- colSums(as.matrix(rawProtDatPat))
normProtDatPat  <- rawProtDatPat[,which(patColSums > 0)]

logApproxProtPat <- datRes[which(fullSce$group == "Pat"),which(patColSums > 0)]
corVecPat <- sapply(seq_along(normProtDatPat), function(x){
    cor(logApproxProtPat[,x], normProtDatPat[,x])
})

names(corVecPat) <- colnames(normProtDatPat)
corVecPat
# CD003_PROT  CD004_PROT  CD008_PROT CD011c_PROT  CD014_PROT  CD016_PROT  CD019_PROT  CD045_PROT  CD056_PROT 
#  0.9517221   0.9551450   0.9200652   0.9632657   0.9791807   0.9365895   0.9145552   0.8030901   0.7662959 

#All perfectly accetable values. 
dir.create("Diagnostics/Estimation_controls/Obs_vs_est_prot/Pat", recursive = TRUE)

for(i in names(corVecPat)){
    locGgDat <- as.data.frame(cbind(normProtDatPat[,i],
                      logApproxProtPat[,i]))
    colnames(locGgDat) <- c("Observed", "Estimated")
    p <- ggplot(locGgDat, aes(x=Observed, y=Estimated) ) +
        geom_bin2d(bins = 70) +
        scale_fill_continuous(type = "viridis") +
        theme_bw() + stat_cor(p.accuracy = 0.001, r.accuracy = 0.01)
    ggsave(paste0("Diagnostics/Estimation_controls/Obs_vs_est_prot/Pat/", i, ".pdf"), plot = p)
}

#Now, the controls
rawProtDatCtrl <- adtCountsFull[which(fullSce$group == "Ctrl"),]
ctrlColSums <- colSums(as.matrix(rawProtDatCtrl))
normProtDatCtrl <- rawProtDatCtrl[,which(ctrlColSums > 0)]

logApproxProtCtrl <- datRes[which(fullSce$group == "Ctrl"),which(ctrlColSums > 0)]
corVecCtrl <- sapply(seq_along(normProtDatCtrl), function(x){
    cor(logApproxProtCtrl[,x], normProtDatCtrl[,x])
})

names(corVecCtrl) <- colnames(normProtDatCtrl)
corVecCtrl[order(corVecCtrl)]

#  CD090_PROT   CD117_PROT   CD080_PROT   CD010_PROT   CD070_PROT   CD103_PROT 
#   0.4237960    0.4440588    0.4833877    0.4964418    0.5005614    0.5679686 
# CX3CR1_PROT   CD279_PROT  CD062L_PROT   CD025_PROT   CD057_PROT   CD034_PROT 
#   0.6642951    0.6947672    0.7196925    0.7269032    0.7613589    0.7694726 
#  CD071_PROT   CD195_PROT    BTLA_PROT   CD069_PROT   CD024_PROT   CD056_PROT 
#   0.7801554    0.7997188    0.7999154    0.8088275    0.8173995    0.8200549 
#  CD123_PROT   CD185_PROT   CD314_PROT  CD001c_PROT   CD163_PROT   CD013_PROT 
#   0.8330743    0.8367982    0.8381819    0.8573970    0.8594802    0.8606219 
#  CD194_PROT   CD278_PROT   KLRG1_PROT HLA.ABC_PROT   CD196_PROT CD045RO_PROT 
#   0.8655482    0.8665317    0.8765441    0.8817847    0.8842169    0.8898482 
#  CD161_PROT   CD141_PROT  CD001d_PROT   CD303_PROT   CD038_PROT   CD127_PROT 
#   0.8922864    0.8942880    0.9001315    0.9002153    0.9057311    0.9070383 
#CD045RA_PROT   CD008_PROT  HLA.DR_PROT   CD040_PROT   CD003_PROT   CD007_PROT 
#   0.9114985    0.9142728    0.9183743    0.9186793    0.9267858    0.9276804 
#  CD086_PROT   CD005_PROT   CD002_PROT   CD039_PROT   CD032_PROT   CD027_PROT 
#   0.9278266    0.9280142    0.9344625    0.9351101    0.9367962    0.9379394 
#    IgD_PROT   CD016_PROT   CD033_PROT   CD020_PROT   CD244_PROT   CD028_PROT 
#   0.9387920    0.9394772    0.9401379    0.9433494    0.9518923    0.9522547 
#  CD018_PROT   CD021_PROT   CD031_PROT   CD004_PROT  CD011b_PROT   CD064_PROT 
#   0.9531578    0.9537869    0.9547734    0.9630434    0.9753844    0.9757622 
# CD011c_PROT   CD019_PROT 
#   0.9788760    0.9800513 

#Interestingly, this works well for the majority, but a fraction of lowly expressed
#proteins get very low correlation scores. These will be investigated one-by one. 
#When having gone through a fraction, it seems like these can be divided in two 
#major groups: proteins that have a scattered expression and proteins that are highly
#expressed by a small group of cells. The latter group seem genereally to do well here
#even if the correlation coefficients are low due to lots of noise signal in the 
#negative fraction. 

dir.create("Diagnostics/Estimation_controls/Obs_vs_est_prot/Ctrl", recursive = TRUE)

for(i in names(corVecCtrl)){
    locGgDat <- as.data.frame(cbind(normProtDatCtrl[,i],
                                    logApproxProtCtrl[,i]))
    colnames(locGgDat) <- c("Observed", "Estimated")
    p <- ggplot(locGgDat, aes(x=Observed, y=Estimated) ) +
        geom_bin2d(bins = 70) +
        scale_fill_continuous(type = "viridis") +
        theme_bw() + stat_cor(p.accuracy = 0.001, r.accuracy = 0.01)
    ggsave(paste0("Diagnostics/Estimation_controls/Obs_vs_est_prot/Ctrl/", i, ".pdf"), plot = p)
}

#Markers that definitely do not work: 
#CD90

#There are in addition a few borderline markers, such as CD70 and CD117

#For these borderline markers, we will make a directed, extra check: 
#Is there a correlation between estimated protein and observed RNA expression? 

gexAdtNameList <- list(c("BTLA_PROT", "BTLA"),
                       c("CD001c_PROT", "CD1C"),
                       c("CD001d_PROT", "CD1D"),
                       c("CD002_PROT", "CD2"),
                       c("CD003_PROT", "CD3E"),
                       c("CD004_PROT", "CD4"),
                       c("CD005_PROT", "CD5"),
                       c("CD007_PROT", "CD7"),
                       c("CD008_PROT", "CD8A"),
                       c("CD010_PROT", "MME"),
                       c("CD011b_PROT", "ITGAM"),
                       c("CD011c_PROT", "ITGAX"),
                       c("CD013_PROT", "ANPEP"),
                       c("CD014_PROT", "CD14"),
                       c("CD016_PROT", "FCGR3A"),
                       c("CD018_PROT", "ITGB2"),
                       c("CD019_PROT", "CD19"),
                       c("CD020_PROT", "MS4A1"),
                       c("CD021_PROT", "CR2"),
                       c("CD024_PROT", "CD24"),
                       c("CD025_PROT", "IL2RA"),
                       c("CD027_PROT", "CD27"),
                       c("CD028_PROT", "CD28"),
                       c("CD031_PROT", "PECAM1"),
                       c("CD032_PROT", "FCGR2A"),
                       c("CD033_PROT", "CD33"),
                       c("CD034_PROT", "CD34"),
                       c("CD038_PROT", "CD38"),
                       c("CD039_PROT", "ENTPD1"),
                       c("CD040_PROT", "CD40"),
                       c("CD045_PROT", "PTPRC"),
                       c("CD045RA_PROT", "PTPRC"),
                       c("CD045RO_PROT", "PTPRC"),
                       c("CD056_PROT", "NCAM1"),
                       c("CD057_PROT", "B3GAT1"),
                       c("CD062L_PROT", "SELL"),
                       c("CD064_PROT", "FCGR1BP"),
                       c("CD069_PROT", "CD69"),
                       c("CD070_PROT", "CD70"),
                       c("CD071_PROT", "TFRC"),
                       c("CD080_PROT", "CD80"),
                       c("CD086_PROT", "CD86"),
                       c("CD090_PROT", "THY1"),
                       c("CD103_PROT", "ITGAE"),
                       c("CD117_PROT", "KIT"),
                       c("CD123_PROT", "IL3RA"),
                       c("CD127_PROT", "IL7R"),
                       c("CD141_PROT", "THBD"),
                       c("CD161_PROT", "KLRB1"),
                       c("CD163_PROT", "CD163"),
                       c("CD185_PROT", "CXCR5"),
                       c("CD194_PROT", "CCR4"),
                       c("CD195_PROT", "CCR5"),
                       c("CD196_PROT", "CCR6"),
                       c("CD244_PROT", "CD244"),
                       c("CD278_PROT", "ICOS"),
                       c("CD279_PROT", "PDCD1"),
                       c("CD303_PROT", "CLEC4C"),
                       c("CD314_PROT", "KLRK1"),
                       c("CX3CR1_PROT", "CX3CR1"),
                       c("HLA.ABC_PROT", "HLA-A"),
                       c("HLA.DR_PROT", "HLA-DRA"),
                       c("IgD_PROT", "IGHD"),
                       c("KLRG1_PROT", "KLRG1"))

adtNameListProt <- unlist(lapply(gexAdtNameList, "[[", 1))
#Here, we exclude the ones that are not preset among the truly expressed proteins
#and that are not present in the RNA space. 

rowDataFullSce <- rowData(fullSce)
adtNameListRna <- unlist(lapply(gexAdtNameList, "[[", 2))
adtNameListRnaPresent <- 
    rowDataFullSce$hgnc_symbol[which(rowDataFullSce$hgnc_symbol %in% adtNameListRna)]
adtNameListRnaNotPresent <- adtNameListRna[-which(adtNameListRna %in% adtNameListRnaPresent )]
adtNameListRnaNotPresent
#"CD24"   "PECAM1" "IGHD" 

gexAdtNameListCtrl <- gexAdtNameList[Reduce(intersect, 
                                            list(which(adtNameListRna %in% 
                                                           adtNameListRnaPresent), 
                                                 which(colnames(datRes) %in% 
                                                           colnames(logApproxProtCtrl))))]

dir.create("Diagnostics/Estimation_controls/RNA_vs_est_prot/Ctrl", recursive = TRUE)
for(i in gexAdtNameListCtrl){
    iProt <- i[1]
    iGene <- i[2]
    locGgDat <- as.data.frame(cbind(logcounts(fullSce)[which(rowDataFullSce$hgnc_symbol == iGene),
                                                       which(fullSce$group == "Ctrl")],
                                    logApproxProtCtrl[,iProt]))
    colnames(locGgDat) <- c("Observed_RNA", "Estimated_protein")
    p <- ggplot(locGgDat, aes(x=Observed_RNA, y=Estimated_protein) ) +
        geom_bin2d(bins = 70) +
        scale_fill_continuous(type = "viridis") +
        theme_bw() + stat_cor(p.accuracy = 0.001, r.accuracy = 0.01)
    ggsave(paste0("Diagnostics/Estimation_controls/RNA_vs_est_prot/Ctrl/", iProt, ".pdf"), plot = p)
}

dir.create("Diagnostics/Estimation_controls/RNA_vs_obs_prot/Ctrl", recursive = TRUE)
for(i in gexAdtNameListCtrl){
    iProt <- i[1]
    iGene <- i[2]
    locGgDat <- as.data.frame(cbind(logcounts(fullSce)[which(rowDataFullSce$hgnc_symbol == iGene),
                                                       which(fullSce$group == "Ctrl")],
                                    normProtDatCtrl[,iProt]))
    colnames(locGgDat) <- c("Observed_RNA", "Observed_protein")
    p <- ggplot(locGgDat, aes(x=Observed_RNA, y=Observed_protein) ) +
        geom_bin2d(bins = 70) +
        scale_fill_continuous(type = "viridis") +
        theme_bw() + stat_cor(p.accuracy = 0.001, r.accuracy = 0.01)
    ggsave(paste0("Diagnostics/Estimation_controls/RNA_vs_obs_prot/Ctrl/", iProt, ".pdf"), plot = p)
}

#And now the patients
gexAdtNameListPat <- gexAdtNameList[Reduce(intersect, 
                                            list(which(adtNameListRna %in% 
                                                           adtNameListRnaPresent), 
                                                 which(colnames(datRes) %in% 
                                                           colnames(logApproxProtPat))))]

dir.create("Diagnostics/Estimation_controls/RNA_vs_est_prot/Pat", recursive = TRUE)
for(i in gexAdtNameListPat){
    iProt <- i[1]
    iGene <- i[2]
    locGgDat <- as.data.frame(cbind(logcounts(fullSce)[which(rowDataFullSce$hgnc_symbol == iGene),
                                                       which(fullSce$group == "Pat")],
                                    logApproxProtPat[,iProt]))
    colnames(locGgDat) <- c("Observed_RNA", "Estimated_protein")
    p <- ggplot(locGgDat, aes(x=Observed_RNA, y=Estimated_protein) ) +
        geom_bin2d(bins = 70) +
        scale_fill_continuous(type = "viridis") +
        theme_bw() + stat_cor(p.accuracy = 0.001, r.accuracy = 0.01)
    ggsave(paste0("Diagnostics/Estimation_controls/RNA_vs_est_prot/Pat/", iProt, ".pdf"), plot = p)
}

dir.create("Diagnostics/Estimation_controls/RNA_vs_obs_prot/Pat", recursive = TRUE)
for(i in gexAdtNameListPat){
    iProt <- i[1]
    iGene <- i[2]
    locGgDat <- as.data.frame(cbind(logcounts(fullSce)[which(rowDataFullSce$hgnc_symbol == iGene),
                                                       which(fullSce$group == "Pat")],
                                    normProtDatPat[,iProt]))
    colnames(locGgDat) <- c("Observed_RNA", "Observed_protein")
    p <- ggplot(locGgDat, aes(x=Observed_RNA, y=Observed_protein) ) +
        geom_bin2d(bins = 70) +
        scale_fill_continuous(type = "viridis") +
        theme_bw() + stat_cor(p.accuracy = 0.001, r.accuracy = 0.01)
    ggsave(paste0("Diagnostics/Estimation_controls/RNA_vs_obs_prot/Pat/", iProt, ".pdf"), plot = p)
}

#The bordserline markers CD56, CD103, CD117 and CD185 all seem to work, judging 
#on the pattern of co-expression which looks quite similar in all four cases.

#We here also investigate two markers present in both datasets,to see which 
#of the two that is preferable, i.e. the original or the estimated. 

pdf("Diagnostics/Estimation_controls/CD3_vs_CD56_observed_vs_approximated.pdf", height = 12, width = 9)
par(mfrow = c(2,2))
plot(logApproxProtCtrl[1:10000,"CD003_PROT"], logApproxProtCtrl[1:10000,"CD056_PROT"], 
     xlab = "CD3", ylab = "CD56", main = "Control_approximated")

plot(normProtDatCtrl[1:10000,"CD003_PROT"], normProtDatCtrl[1:10000,"CD056_PROT"],
     xlab = "CD3", ylab = "CD56", main = "Control_observed")

plot(logApproxProtPat[1:10000,"CD003_PROT"], logApproxProtPat[1:10000,"CD056_PROT"],
     xlab = "CD3", ylab = "CD56", main = "Patient_approximated")

plot(normProtDatPat[1:10000,"CD003_PROT"], normProtDatPat[1:10000,"CD056_PROT"], 
     xlab = "CD3", ylab = "CD56", main = "Patient_observed")
dev.off()

#It is clear from this, that the estimates both give better separation and are more
#similar to each other. There is strange, middle population that is positive for both
#CD56 and CD3 which seems to originate from the patient data, but that will be clearly
#separated in the clustering anyhow, so it should not be a problem for downstream analyses. 

#CD10, CD70 and CD90 are now excluded and the erstimates are kept in the other
#instances. 


altExp(fullSce) <- altExp(fullSce)[-which(row.names(altExp(fullSce)) == "CD090_PROT"),]

dim(altExp(fullSce)) #63
assays(altExp(fullSce)) #counts normcounts logcounts

#So here we are, transformed and ready for the coming steps. 
#And here, the full dataset is saved
saveRDS(fullSce, "../External/Stockholm/Data/SCE_versions/2_normalised_SCE_with_full_protein.rds")

#And we also save the three datasets for visualisation purposes
write.csv(adtCountsCtrl, "Data/Stockholm/totalVI_protein_corrections/adtCountsCtrl.csv")
write.csv(normAdtCountsCtrl, "Data/Stockholm/totalVI_protein_corrections/normAdtCountsCtrl.csv")
write.csv(transformed_x, "Data/Stockholm/totalVI_protein_corrections/transformed_x.csv")
